gamma comparison ongoing

+++ diffusers_implementation/config.py
@@ -5,8 +5,8 @@ def get_args():
     parse = argparse.ArgumentParser()
     
     # hyperparameter of StyleID
-    parse.add_argument('--T', type=int, default=1.5)
-    parse.add_argument('--gamma', type=int, default=0.75)
+    parse.add_argument('--T', type=float, default=1.5)
+    parse.add_argument('--gamma', type=float, default=0.75)
     parse.add_argument('--without_init_adain', action='store_true')
     parse.add_argument('--without_attn_injection', action='store_true')
     parse.add_argument('--layers', nargs='+', type=int, default=[7, 8, 9, 10, 11])
